# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AL JUDE Travel — Internal Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  /** @nullable */
  displayName?: string | null;
}

export interface PageInfo {
  slug: string;
  title: string;
  titleAr: string;
  icon: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}